Rename constraint_solver into local_constraint_solver
