+++ moveit_ros/hybrid_planning/local_planner/src/local_constraint_solver_interface.cpp
@@ -32,56 +32,9 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-/* Author: Sebastian Jahr
-   Description: Defines an interface for a local constraint solver plugin implementation for the local planner component node.
- */
-
-#pragma once
-
-#include <rclcpp/rclcpp.hpp>
-#include <rclcpp_action/rclcpp_action.hpp>
-
-#include <moveit/planning_scene_monitor/planning_scene_monitor.h>
-#include <moveit/robot_state/robot_state.h>
-
-#include <moveit_msgs/msg/robot_trajectory.hpp>
-#include <moveit_msgs/msg/constraints.hpp>
-
-#include <moveit_msgs/action/local_planner.hpp>
-
-#include <trajectory_msgs/msg/joint_trajectory.h>
+#include <moveit/local_planner/local_constraint_solver_interface.h>
 
 namespace moveit_hybrid_planning
 {
-/**
- * Class ConstraintSolverInterface - Base class for a local constrain solver.
- */
-class ConstraintSolverInterface
-{
-public:
-  /**
-   * Initialize constraint solver
-   * @return True if initialization was successful
-   */
-  virtual bool initialize(const rclcpp::Node::SharedPtr& node,
-                          planning_scene_monitor::PlanningSceneMonitorPtr planning_scene_monitor) = 0;
-
-  /**
-   * Solve local planning problem for the current loop run
-   * @param local_trajectory The local trajectory to pursue
-   * @param local_constraints Local goal constraints
-   * @param planning_scene The planning scene to use for local planning
-   * @param feedback Feedback event string from the current solver i.e. "Collision detected"
-   * @return Local planning solution in joint space
-   */
-  virtual trajectory_msgs::msg::JointTrajectory
-  solve(robot_trajectory::RobotTrajectory local_trajectory,
-        std::vector<moveit_msgs::msg::Constraints> local_constraints,
-        std::shared_ptr<moveit_msgs::action::LocalPlanner::Feedback> feedback) = 0;
-  virtual ~ConstraintSolverInterface(){};
-
-protected:
-  /** \brief Constructor */
-  ConstraintSolverInterface(){};
-};
-}  // namespace moveit_hybrid_planning
+// Empty because this library only defines an interface
+}
